+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `[Datepicker]` Added support for custom api callback to disable passed dates and to disable dates by years. ([#3462](https://github.com/infor-design/enterprise/issues/3462))
 - `[Contextmenu]` Added support for shortcut display in menus. ([#3490](https://github.com/infor-design/enterprise/issues/3490))
+- `[EmptyMessage]` Added a new design for custom error message. ([#281](https://github.com/infor-design/design-system/issues/281))
 
 ### v4.28.0 Fixes